crawler/requirements.txt
playwright==1.42.0
google-genai>=0.1.0
pydantic==2.11.1
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
httpx>=0.27.0
pgvector>=0.2.4
python-dotenv>=1.0.0
